+++ src/token/signature.h
@@ -17,8 +17,14 @@
  *   Pointer to token information to be signed.
  * @param[out] signature
  *   Pointer to signature byte buffer.
+ * @param[in]  siglen
+ *   Length of signature buffer.
+ * @return 0 if success
  */
-void sign_token(uint8_t *secret, token_t *token, uint8_t *signature);
+int sign_token(uint8_t secret[static SECRET_LEN],
+               token_t *token,
+               uint8_t *signature,
+               size_t siglen);
 
 /**
  * Verify a token signature
@@ -29,11 +35,13 @@ void sign_token(uint8_t *secret, token_t *token, uint8_t *signature);
  *   Pointer to token information to be signed.
  * @param[in] signature
  *   Pointer to signature byte buffer.
+ * @param[in]  siglen
+ *   Length of signature buffer.
  *
- * @return bool if signature is valid of not.
+ * @return int 0 if signature is valid, 1 if invalid, anything else is an error.
  *
  */
-bool verify_token_signature(uint8_t *secret, token_t *token, uint8_t *signature);
+int verify_token_signature(uint8_t *secret, token_t *token, uint8_t *signature, size_t siglen);
 
 /**
  * Verify a token signature from the raw apdu data
@@ -43,7 +51,8 @@ bool verify_token_signature(uint8_t *secret, token_t *token, uint8_t *signature)
  * @param[out] token
  *   Token parsed.
  *
- * @return bool if signature is valid of not.
+ * @return int 0 if signature is valid, 1 if invalid, anything else is an error.
+ *
  *
  */
-bool check_token_signature_from_apdu(buffer_t *cdata, token_t *token);
+int check_token_signature_from_apdu(buffer_t *cdata, token_t *token);